requirements.txt
Django
django-bootstrap4
django-bootstrap-datepicker-plus
Pillow
django-crontab
djangorestframework
djangorestframework-simplejwt
django-cors-headers
